Pet Growth Alteration
In the original implementation, the pet grew by a scale of 1 each time it was fed. Due to the model being at a default of 20 and the max being 27. This did not seem adequate enough to make the growth feel gradual. Thus, the scale alteration was changed to .25 instead of one making the growth occur over a longer span of time.